pyproject.toml
[project]
name = "pycg"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo-Blois",email = "gustavobloisrr@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyopengl (>=3.1.9,<4.0.0)",
    "glfw (>=2.9.0,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "glm (>=0.4.4,<0.5.0)",
    "pyglm (>=2.8.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
